+++ src/RequestDecoder/JsonRequestDecoder.php
@@ -6,8 +6,9 @@ namespace DigitalCraftsman\CQSRouting\RequestDecoder;
 
 use Symfony\Component\HttpFoundation\Request;
 
-final readonly class JsonRequestDecoder implements RequestDecoderInterface
+final readonly class JsonRequestDecoder implements RequestDecoder
 {
+    #[\Override]
     public function decodeRequest(Request $request): array
     {
         /**