+++ src/ResponseConstructor/EmptyJsonResponseConstructor.php
@@ -8,8 +8,9 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-final readonly class EmptyJsonResponseConstructor implements ResponseConstructorInterface
+final readonly class EmptyJsonResponseConstructor implements ResponseConstructor
 {
+    #[\Override]
     public function constructResponse($data, Request $request): JsonResponse
     {
         return new JsonResponse('', Response::HTTP_NO_CONTENT, [], true);