+++ src/ResponseConstructor/EmptyResponseConstructor.php
@@ -7,8 +7,9 @@ namespace DigitalCraftsman\CQSRouting\ResponseConstructor;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-final readonly class EmptyResponseConstructor implements ResponseConstructorInterface
+final readonly class EmptyResponseConstructor implements ResponseConstructor
 {
+    #[\Override]
     public function constructResponse($data, Request $request): Response
     {
         return new Response(null, Response::HTTP_NO_CONTENT);