+++ src/ResponseConstructor/StreamedResponseConstructor.php
@@ -8,11 +8,12 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 
-final readonly class StreamedResponseConstructor implements ResponseConstructorInterface
+final readonly class StreamedResponseConstructor implements ResponseConstructor
 {
     /**
      * @param callable $data
      */
+    #[\Override]
     public function constructResponse($data, Request $request): Response
     {
         return new StreamedResponse($data);