+++ tests/Test/Application/AddActionIdRequestDataTransformer.php
@@ -6,15 +6,16 @@ namespace DigitalCraftsman\CQSRouting\Test\Application;
 
 use DigitalCraftsman\CQSRouting\Command\Command;
 use DigitalCraftsman\CQSRouting\Query\Query;
-use DigitalCraftsman\CQSRouting\RequestDataTransformer\RequestDataTransformerInterface;
+use DigitalCraftsman\CQSRouting\RequestDataTransformer\RequestDataTransformer;
 use DigitalCraftsman\CQSRouting\Test\ValueObject\ActionId;
 
-final class AddActionIdRequestDataTransformer implements RequestDataTransformerInterface
+final class AddActionIdRequestDataTransformer implements RequestDataTransformer
 {
     /**
      * @param class-string<Command|Query> $dtoClass
      * @param null                        $parameters
      */
+    #[\Override]
     public function transformRequestData(
         string $dtoClass,
         array $requestData,
@@ -26,6 +27,7 @@ final class AddActionIdRequestDataTransformer implements RequestDataTransformerI
     }
 
     /** @param null $parameters */
+    #[\Override]
     public static function areParametersValid(mixed $parameters): bool
     {
         return $parameters === null;