+++ tests/Test/Application/ConnectionTransactionWrapper.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace DigitalCraftsman\CQSRouting\Test\Application;
 
 use DigitalCraftsman\CQSRouting\Command\Command;
-use DigitalCraftsman\CQSRouting\HandlerWrapper\HandlerWrapperInterface;
+use DigitalCraftsman\CQSRouting\HandlerWrapper\HandlerWrapper;
 use DigitalCraftsman\CQSRouting\Query\Query;
 use DigitalCraftsman\CQSRouting\Test\Utility\ConnectionSimulator;
 use Symfony\Component\HttpFoundation\Request;
 
-final class ConnectionTransactionWrapper implements HandlerWrapperInterface
+final class ConnectionTransactionWrapper implements HandlerWrapper
 {
     public function __construct(
         private readonly ConnectionSimulator $connectionSimulator,
@@ -18,6 +18,7 @@ final class ConnectionTransactionWrapper implements HandlerWrapperInterface
     }
 
     /** @param null $parameters */
+    #[\Override]
     public function prepare(
         Command | Query $dto,
         Request $request,
@@ -27,6 +28,7 @@ final class ConnectionTransactionWrapper implements HandlerWrapperInterface
     }
 
     /** @param null $parameters */
+    #[\Override]
     public function catch(
         Command | Query $dto,
         Request $request,
@@ -39,6 +41,7 @@ final class ConnectionTransactionWrapper implements HandlerWrapperInterface
     }
 
     /** @param null $parameters */
+    #[\Override]
     public function then(
         Command | Query $dto,
         Request $request,
@@ -49,22 +52,26 @@ final class ConnectionTransactionWrapper implements HandlerWrapperInterface
 
     // Priorities
 
+    #[\Override]
     public static function preparePriority(): int
     {
         return 50;
     }
 
+    #[\Override]
     public static function catchPriority(): int
     {
         return 50;
     }
 
+    #[\Override]
     public static function thenPriority(): int
     {
         return 50;
     }
 
     /** @param null $parameters */
+    #[\Override]
     public static function areParametersValid(mixed $parameters): bool
     {
         return $parameters === null;