+++ tests/Test/Application/UserIdValidator.php
@@ -5,16 +5,16 @@ declare(strict_types=1);
 namespace DigitalCraftsman\CQSRouting\Test\Application;
 
 use DigitalCraftsman\CQSRouting\Command\Command;
-use DigitalCraftsman\CQSRouting\DTOValidator\DTOValidatorInterface;
+use DigitalCraftsman\CQSRouting\DTOValidator\DTOValidator;
 use DigitalCraftsman\CQSRouting\Query\Query;
 use DigitalCraftsman\CQSRouting\Test\Domain\News\WriteSide\CreateNewsArticle\CreateNewsArticleCommand;
 use DigitalCraftsman\CQSRouting\Test\Utility\SecuritySimulator;
 use Symfony\Component\HttpFoundation\Request;
 
-final class UserIdValidator implements DTOValidatorInterface
+final readonly class UserIdValidator implements DTOValidator
 {
     public function __construct(
-        private readonly SecuritySimulator $securitySimulator,
+        private SecuritySimulator $securitySimulator,
     ) {
     }
 
@@ -22,6 +22,7 @@ final class UserIdValidator implements DTOValidatorInterface
      * @param CreateNewsArticleCommand $dto
      * @param null                     $parameters
      */
+    #[\Override]
     public function validateDTO(
         Request $request,
         Command | Query $dto,
@@ -33,6 +34,7 @@ final class UserIdValidator implements DTOValidatorInterface
     }
 
     /** @param null $parameters */
+    #[\Override]
     public static function areParametersValid(mixed $parameters): bool
     {
         return $parameters === null;