+++ tests/Test/Domain/News/WriteSide/CreateNewsArticle/FailingCreateNewsArticleCommandHandler.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\CQSRouting\Test\Domain\News\WriteSide\CreateNewsArticle;
 
-use DigitalCraftsman\CQSRouting\Command\CommandHandlerInterface;
+use DigitalCraftsman\CQSRouting\Command\CommandHandler;
 use DigitalCraftsman\CQSRouting\Test\Domain\News\WriteSide\CreateNewsArticle\Exception\NewsArticleAlreadyExists;
 
-final readonly class FailingCreateNewsArticleCommandHandler implements CommandHandlerInterface
+final readonly class FailingCreateNewsArticleCommandHandler implements CommandHandler
 {
     public function __invoke(CreateNewsArticleCommand $command): void
     {