+++ tests/Test/Domain/Tasks/ReadSide/GetTasks/GetTasksQueryHandler.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\CQSRouting\Test\Domain\Tasks\ReadSide\GetTasks;
 
-use DigitalCraftsman\CQSRouting\Query\QueryHandlerInterface;
+use DigitalCraftsman\CQSRouting\Query\QueryHandler;
 use DigitalCraftsman\CQSRouting\Test\Repository\TasksInMemoryRepository;
 
-final readonly class GetTasksQueryHandler implements QueryHandlerInterface
+final readonly class GetTasksQueryHandler implements QueryHandler
 {
     public function __construct(
         private TasksInMemoryRepository $tasksInMemoryRepository,