+++ tests/Test/Domain/Tasks/WriteSide/CreateTask/CreateTaskRequestDecoder.php
@@ -4,11 +4,12 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\CQSRouting\Test\Domain\Tasks\WriteSide\CreateTask;
 
-use DigitalCraftsman\CQSRouting\RequestDecoder\RequestDecoderInterface;
+use DigitalCraftsman\CQSRouting\RequestDecoder\RequestDecoder;
 use Symfony\Component\HttpFoundation\Request;
 
-final class CreateTaskRequestDecoder implements RequestDecoderInterface
+final class CreateTaskRequestDecoder implements RequestDecoder
 {
+    #[\Override]
     public function decodeRequest(Request $request): array
     {
         return [