+++ tests/Test/Domain/Tasks/WriteSide/DefineTaskHourContingent/DefineTaskHourContingentRequestDataTransformer.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\CQSRouting\Test\Domain\Tasks\WriteSide\DefineTaskHourContingent;
 
-use DigitalCraftsman\CQSRouting\RequestDataTransformer\RequestDataTransformerInterface;
+use DigitalCraftsman\CQSRouting\RequestDataTransformer\RequestDataTransformer;
 
-final class DefineTaskHourContingentRequestDataTransformer implements RequestDataTransformerInterface
+final class DefineTaskHourContingentRequestDataTransformer implements RequestDataTransformer
 {
     /**
      * @param class-string $dtoClass
@@ -15,6 +15,7 @@ final class DefineTaskHourContingentRequestDataTransformer implements RequestDat
      * } $requestData
      * @param null $parameters
      */
+    #[\Override]
     public function transformRequestData(string $dtoClass, array $requestData, mixed $parameters): array
     {
         $requestData['hourContingent'] = (float) $requestData['hourContingent'];
@@ -23,6 +24,7 @@ final class DefineTaskHourContingentRequestDataTransformer implements RequestDat
     }
 
     /** @param null $parameters */
+    #[\Override]
     public static function areParametersValid(mixed $parameters): bool
     {
         return $parameters === null;